Fix #9 Supplier references collisions lead to wrong supplier price
The old code used supplier references associated in conjunction with fk_product.
This triggers a database search.
Since supplier references can collide, we may not get the price we want.

This fix uses the price ID directly, so we control the value we get.
This is done at the cost of some ugliness of the code.